move more funcs
