+++ iroh-net/src/hp/netcheck.rs
@@ -21,9 +21,8 @@ use crate::net::ip::to_canonical;
 
 use self::probe::{Probe, ProbeProto};
 
-use super::derp::{DerpMap, DerpNode, DerpRegion, UseIpv4, UseIpv6};
+use super::derp::{DerpMap, DerpNode, UseIpv4, UseIpv6};
 use super::dns::DNS_RESOLVER;
-use super::ping::Pinger;
 use super::portmapper;
 use super::stun;
 
@@ -256,72 +255,6 @@ impl Client {
             Err(_) => Err(anyhow!("channel closed, actor awol")),
         }
     }
-}
-
-async fn measure_https_latency(_reg: &DerpRegion) -> Result<(Duration, IpAddr)> {
-    anyhow::bail!("not implemented");
-    // TODO:
-    // - needs derphttp::Client
-    // - measurement hooks to measure server processing time
-
-    // metricHTTPSend.Add(1)
-    // let ctx, cancel := context.WithTimeout(httpstat.WithHTTPStat(ctx, &result), overallProbeTimeout);
-    // let dc := derphttp.NewNetcheckClient(c.logf);
-    // let tlsConn, tcpConn, node := dc.DialRegionTLS(ctx, reg)?;
-    // if ta, ok := tlsConn.RemoteAddr().(*net.TCPAddr);
-    // req, err := http.NewRequestWithContext(ctx, "GET", "https://"+node.HostName+"/derp/latency-check", nil);
-    // resp, err := hc.Do(req);
-
-    // // DERPs should give us a nominal status code, so anything else is probably
-    // // an access denied by a MITM proxy (or at the very least a signal not to
-    // // trust this latency check).
-    // if resp.StatusCode > 299 {
-    //     return 0, ip, fmt.Errorf("unexpected status code: %d (%s)", resp.StatusCode, resp.Status)
-    // }
-    // _, err = io.Copy(io.Discard, io.LimitReader(resp.Body, 8<<10));
-    // result.End(c.timeNow())
-
-    // // TODO: decide best timing heuristic here.
-    // // Maybe the server should return the tcpinfo_rtt?
-    // return result.ServerProcessing, ip, nil
-}
-
-async fn measure_icmp_latency(reg: &DerpRegion, p: &Pinger) -> Result<Duration> {
-    if reg.nodes.is_empty() {
-        anyhow::bail!(
-            "no nodes for region {} ({})",
-            reg.region_id,
-            reg.region_code
-        );
-    }
-
-    // Try pinging the first node in the region
-    let node = &reg.nodes[0];
-
-    // Get the IPAddr by asking for the UDP address that we would use for
-    // STUN and then using that IP.
-    let node_addr = get_derp_addr(node, ProbeProto::Ipv4)
-        .await
-        .with_context(|| format!("no address for node {}", node.name))?;
-
-    debug!(
-        "ICMP ping start to {} with payload len {} - derp {} {}",
-        node_addr,
-        node.name.as_bytes().len(),
-        node.name,
-        reg.region_id
-    );
-    // Use the unique node.name field as the packet data to reduce the
-    // likelihood that we get a mismatched echo response.
-    let d = p.send(node_addr.ip(), node.name.as_bytes()).await?;
-    debug!(
-        "ICMP ping done {} with latency {}ms - derp {} {}",
-        node_addr,
-        d.as_millis(),
-        node.name,
-        reg.region_id
-    );
-    Ok(d)
 }
 
 /// Returns the IP address to use to communicate to this derp node.
@@ -991,6 +924,7 @@ mod tests {
     use bytes::BytesMut;
     use tokio::time;
 
+    use crate::hp::derp::DerpRegion;
     use crate::test_utils::setup_logging;
 
     use super::*;
